+++ src/bulkUtils.ts
@@ -5,14 +5,22 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import type {
+import { Transform, Readable } from 'node:stream';
+import { pipeline } from 'node:stream/promises';
+import * as fs from 'node:fs';
+import { EOL } from 'node:os';
+import { HttpApi } from '@jsforce/jsforce-node/lib/http-api.js';
+import { HttpResponse } from '@jsforce/jsforce-node';
+import {
   JobInfoV2,
   IngestJobV2Results,
   IngestJobV2SuccessfulResults,
   IngestJobV2FailedResults,
   IngestJobV2UnprocessedRecords,
+  QueryJobV2,
+  QueryJobInfoV2,
 } from '@jsforce/jsforce-node/lib/api/bulk2.js';
-
+import { Parser as csvParse } from 'csv-parse';
 import type { Schema } from '@jsforce/jsforce-node';
 import { Connection, Messages } from '@salesforce/core';
 import { IngestJobV2 } from '@jsforce/jsforce-node/lib/api/bulk2.js';
@@ -174,3 +182,127 @@ export const remainingTime =
   (now: number) =>
   (endWaitTime?: number): number =>
     Math.max((endWaitTime ?? now) - now, 0);
+
+export enum ColumnDelimiter {
+  BACKQUOTE = '`',
+  CARET = '^',
+  COMMA = ',',
+  PIPE = '|',
+  SEMICOLON = ';',
+  TAB = '	',
+}
+
+export type ColumnDelimiterKeys = keyof typeof ColumnDelimiter;
+
+async function bulkRequest(conn: Connection, url: string): Promise<{ body: string; headers: HttpResponse['headers'] }> {
+  const httpApi = new HttpApi(conn, {
+    responseType: 'text/plain', // this ensures jsforce doesn't try parsing the body
+  });
+
+  let headers: HttpResponse['headers'] | undefined;
+
+  httpApi.on('response', (response: HttpResponse) => {
+    headers = response.headers;
+  });
+
+  const body = await httpApi.request<string>({
+    url: conn.normalizeUrl(url),
+    method: 'GET',
+  });
+
+  if (!headers) throw new Error('failed to get HTTP headers for bulk query');
+
+  return {
+    body,
+    headers,
+  };
+}
+
+export async function exportRecords(
+  conn: Connection,
+  queryJob: QueryJobV2<Schema>,
+  outputInfo: {
+    filePath: string;
+    format: 'csv' | 'json';
+    columnDelimiter: ColumnDelimiterKeys;
+  }
+): Promise<QueryJobInfoV2> {
+  let jobInfo: QueryJobInfoV2 | undefined;
+
+  queryJob.on('jobComplete', (completedJob: QueryJobInfoV2) => {
+    jobInfo = completedJob;
+  });
+
+  await queryJob.poll();
+
+  if (jobInfo === undefined) {
+    throw new Error('could not get job info after polling');
+  }
+
+  let locator: string | undefined;
+
+  let recordsWritten = 0;
+
+  while (locator !== 'null') {
+    // we can't parallelize this because we:
+    // 1. need to get 1 batch to know the locator for the next one
+    // 2. merge all batches into one csv or json file
+    //
+    // eslint-disable-next-line no-await-in-loop
+    const res = await bulkRequest(
+      conn,
+      locator ? `/jobs/query/${jobInfo.id}/results?locator=${locator}` : `/jobs/query/${jobInfo.id}/results`
+    );
+
+    if (outputInfo.format === 'json') {
+      const jsonWritable = fs.createWriteStream(outputInfo.filePath, {
+        // Open file for reading and appending. The file is created if it does not exist but fails if the path exists.
+        // https://nodejs.org/api/fs.html#file-system-flags
+        flags: 'ax+',
+      });
+
+      const totalRecords = jobInfo.numberRecordsProcessed;
+
+      if (!locator) {
+        // first write, start JSON array
+        jsonWritable.write(`[${EOL}`);
+      }
+
+      // eslint-disable-next-line no-await-in-loop
+      await pipeline(
+        Readable.from(res.body),
+        new csvParse({ columns: true, delimiter: ColumnDelimiter[outputInfo.columnDelimiter] }),
+        new Transform({
+          objectMode: true,
+          // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
+          transform(chunk, _encoding, callback) {
+            if (recordsWritten === totalRecords - 1) {
+              callback(null, `  ${JSON.stringify(chunk)}${EOL}]`);
+            } else {
+              recordsWritten++;
+              callback(null, `  ${JSON.stringify(chunk)},${EOL}`);
+            }
+          },
+        }),
+        jsonWritable
+      );
+    } else {
+      // csv
+      // eslint-disable-next-line no-await-in-loop
+      await pipeline(
+        locator
+          ? [
+              Readable.from(res.body.slice(res.body.indexOf(EOL) + 1)),
+              fs.createWriteStream(outputInfo.filePath, {
+                flags: 'a', // append mode
+              }),
+            ]
+          : [Readable.from(res.body), fs.createWriteStream(outputInfo.filePath)]
+      );
+    }
+
+    locator = res.headers['sforce-locator'];
+  }
+
+  return jobInfo;
+}